+++ src/lib/pages/account-details/components/tables/instantiated-contracts/index.tsx
@@ -1,19 +1,16 @@
 import { observer } from "mobx-react-lite";
 
-import { useTierConfig } from "lib/app-provider";
+import { TierSwitcher } from "lib/components/TierSwitcher";
 
 import { InstantiatedContractsTableFull } from "./Full";
 import { InstantiatedContractsTableLite } from "./Lite";
 import type { InstantiatedContractsTableProps } from "./types";
 
 export const InstantiatedContractsTable = observer(
-  (props: InstantiatedContractsTableProps) => {
-    const isFullTier = useTierConfig() === "full";
-
-    return isFullTier ? (
-      <InstantiatedContractsTableFull {...props} />
-    ) : (
-      <InstantiatedContractsTableLite {...props} />
-    );
-  }
+  (props: InstantiatedContractsTableProps) => (
+    <TierSwitcher
+      full={<InstantiatedContractsTableFull {...props} />}
+      lite={<InstantiatedContractsTableLite {...props} />}
+    />
+  )
 );