+++ src/lib/pages/code-details/index.tsx
@@ -25,6 +25,7 @@ import { VerifyPublishCodeModal } from "lib/components/modal";
 import PageContainer from "lib/components/PageContainer";
 import { CelatoneSeo } from "lib/components/Seo";
 import { ErrorFetching, InvalidState } from "lib/components/state";
+import { TierSwitcher } from "lib/components/TierSwitcher";
 import { UserDocsLink } from "lib/components/UserDocsLink";
 import { useSchemaStore } from "lib/providers/store";
 import { useCodeData } from "lib/services/wasm/code";
@@ -50,8 +51,9 @@ const InvalidCode = () => <InvalidState title="Code does not exist" />;
 
 const CodeDetailsBody = observer(({ codeId, tab }: CodeDetailsBodyProps) => {
   const isMobile = useMobile();
-  const isFullTier = useTierConfig() === "full";
   const { isOpen, onClose, onOpen } = useDisclosure();
+  const { isFullTier } = useTierConfig();
+
   const navigate = useInternalNavigate();
   const { getSchemaByCodeHash } = useSchemaStore();
 
@@ -135,11 +137,11 @@ const CodeDetailsBody = observer(({ codeId, tab }: CodeDetailsBodyProps) => {
               attached={!!jsonSchema}
               toJsonSchemaTab={handleTabChange(TabIndex.JsonSchema)}
             />
-            {isFullTier ? (
-              <CodeContractsTableFull codeId={codeId} />
-            ) : (
-              <CodeContractsTableLite codeId={codeId} />
-            )}
+            <TierSwitcher
+              full={<CodeContractsTableFull codeId={codeId} />}
+              lite={<CodeContractsTableLite codeId={codeId} />}
+            />
+
             <UserDocsLink
               title="What is Code in CosmWasm?"
               cta="Read more about Code Details"