+++ CHANGELOG.md
@@ -40,10 +40,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Features
 
 - [#1002](https://github.com/alleslabs/celatone-frontend/pull/1002) Add contract verification modal
+- [#994](https://github.com/alleslabs/celatone-frontend/pull/994) Add Sequencer, Mesa tier and TierSwitcher component
 
 ### Improvements
 
 ### Bug fixes
+
+- [#1004](https://github.com/alleslabs/celatone-frontend/pull/1004) Fix empty string moniker
 
 ## v1.7.0